samsung_epaper/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
numpy>=1.24.0
croniter>=1.3.0
python-multipart>=0.0.6
python-samsung-mdc>=1.15.0
openai>=1.0.0
python-dotenv>=1.0.0
